introduce code formatting and cleanup, copyright notice on all sources.

+++ core/src/main/java/org/tdmx/client/crypto/converters/NumberToOctetString.java
@@ -1,3 +1,21 @@
+/*
+ * TDMX - Trusted Domain Messaging eXchange
+ * 
+ * Enterprise B2B messaging between separate corporations via interoperable cloud service providers.
+ * 
+ * Copyright (C) 2014 Peter Klauser (http://tdmx.org)
+ * 
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General
+ * Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any
+ * later version.
+ * 
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
+ * details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License along with this program. If not, see
+ * http://www.gnu.org/licenses/.
+ */
 package org.tdmx.client.crypto.converters;
 
 import java.nio.ByteBuffer;
@@ -5,23 +23,23 @@ import java.nio.ByteBuffer;
 public class NumberToOctetString {
 
 	public static byte[] intToByte(int x) {
-		return new byte[] {(byte)x};
+		return new byte[] { (byte) x };
 	}
 
 	public static int byteToInt(byte[] bytes) {
-		return (int)bytes[0];
+		return bytes[0];
 	}
 
 	public static byte[] longToBytes(long x) {
-	    ByteBuffer buffer = ByteBuffer.allocate(8);
-	    buffer.putLong(x);
-	    return buffer.array();
+		ByteBuffer buffer = ByteBuffer.allocate(8);
+		buffer.putLong(x);
+		return buffer.array();
 	}
 
 	public static long bytesToLong(byte[] bytes) {
-	    ByteBuffer buffer = ByteBuffer.allocate(8);
-	    buffer.put(bytes);
-	    buffer.flip();//need flip 
-	    return buffer.getLong();
+		ByteBuffer buffer = ByteBuffer.allocate(8);
+		buffer.put(bytes);
+		buffer.flip();// need flip
+		return buffer.getLong();
 	}
 }
